extract class
